rewrite of logic to create signature for emergency upgrade.
