+++ apps/web/app/common/user-role-schema.ts
@@ -1,4 +1,4 @@
 import { z } from "zod";
 
-export const UserRoleSchema = z.enum(["guardian", "securityCouncil", "zkFoundation", "visitor"]);
-export type UserRole = z.infer<typeof UserRoleSchema>;
+export const UserRoleEnum = z.enum(["guardian", "securityCouncil", "zkFoundation", "visitor"]);
+export type UserRole = z.infer<typeof UserRoleEnum>;
